simulator/device_sim/requirements.txt
paho-mqtt==1.6.1
python-dateutil==2.9.0.post0
